requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
azt-connect==1.1.38
billiard==3.6.4.0
boto3==1.18.46
botocore==1.21.46
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
Deprecated==1.2.13
dj-storage==0.2.0
Django==3.2.7
django-cors-headers==3.8.0
django-filter==21.1
django-oauth-toolkit==1.5.0
django-phonenumber-field==5.2.0
django-phonenumbers==1.0.1
django-s3direct==1.1.8
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
geoip2==4.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==1.0
kiwisolver==1.3.2
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.31
Pillow==8.3.2
pip-chill==1.0.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
pygeoip==0.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
transliterate==1.10.2
typing-extensions==3.10.0.2
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
